Change QQuickEventPoint::pointId to be plain int
After change 8984c42d1779b13fd29d95274af2d01d32528e52 in qtbase, a
QTouchEvent::TouchPoint::id is already guaranteed to be unique across
devices. We therefore don't need a larger value space for
QQuickEventPoint::pointId, since that value is the same value as we got
from the QTouchEvent

+++ src/quick/handlers/qquickmultipointerhandler.cpp
@@ -119,7 +119,7 @@ bool QQuickMultiPointerHandler::sameAsCurrentPoints(QQuickPointerEvent *event)
     // or use std::equal with a predicate
     for (int i = 0; ret && i < c; ++i) {
         bool found = false;
-        quint64 pointId = event->point(i)->pointId();
+        int pointId = event->point(i)->pointId();
         for (QQuickEventPoint *o : qAsConst(m_currentPoints))
             if (o && pointId == o->pointId())
                 found = true;
